Move more core-related stuff into Core.cpp/h
