+++ Core/MIPS/ARM/ArmAsm.cpp
@@ -20,7 +20,7 @@
 
 #include "Core/MemMap.h"
 #include "Core/MIPS/MIPS.h"
-#include "Core/System.h"
+#include "Core/Core.h"
 #include "Core/CoreTiming.h"
 #include "Common/MemoryUtil.h"
 #include "Common/CPUDetect.h"
