+++ Core/MIPS/RiscV/RiscVAsm.cpp
@@ -22,7 +22,7 @@
 #include "Core/MIPS/RiscV/RiscVRegCache.h"
 #include "Core/MIPS/JitCommon/JitCommon.h"
 #include "Core/MIPS/JitCommon/JitState.h"
-#include "Core/System.h"
+#include "Core/Core.h"
 
 namespace MIPSComp {
