+++ Core/MIPS/x86/X64IRAsm.cpp
@@ -25,7 +25,7 @@
 #include "Core/MIPS/x86/X64IRRegCache.h"
 #include "Core/MIPS/JitCommon/JitCommon.h"
 #include "Core/MIPS/JitCommon/JitState.h"
-#include "Core/System.h"
+#include "Core/Core.h"
 
 namespace MIPSComp {
